Fix #291.
Update DrawString to use doubles for the position of the string instead of integers.  The use
of doubles allows two things:
* subpixel positioning accuracy for renderers which support AA text rendering
* support for device spaces which are not pixel-based


git-svn-id: http://svn.osgeo.org/mapguide/trunk/MgDev@2666 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ Common/Renderers/AGGRenderer.h
@@ -154,8 +154,8 @@ public:
     RENDERERS_API void SetWorldToScreenTransform(SE_Matrix& xform);
 
     void DrawString(const RS_String& s,
-                    int              x,
-                    int              y,
+                    double           x,
+                    double           y,
                     double           width,
                     double           height,
                     const RS_Font*   font,
@@ -210,12 +210,12 @@ public:
 
     static void DrawString(agg_context*     cxt,
                            const RS_String& s,
-                           int              x,
-                           int              y,
+                           double           x,
+                           double           y,
                            double           width,
                            double           height,
                            const RS_Font*   font,
-                           RS_Color&  color,
+                           RS_Color&        color,
                            double           angle);
